requirements.txt
streamlit_code_editor
lxml
python-evtx
streamlit
lief